Update all java.util.Date references to java8 java.time.ZonedDateTime
Use the new time API to define dates and times with a specific TimeZone (use UTC as default).
Bump Hibernate to version 5, the first providing support to java8 time types.
Add needed dependencies and upgrade Spring to latest version.
Use hibernate5 classes for transactions and sessions in spring files.
Define the useName attribute in EnumType properties for hibernate mappings, as needed for version 5.
For additional information on migration to Hibernate 5 see
https://github.com/hibernate/hibernate-orm/blob/5.0/migration-guide.adoc